% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/counts2fpkm_htseq.R
\name{counts2fpkm_htseq}
\alias{counts2fpkm_htseq}
\title{From counts to fpkm and quartiles with htseq files}
\usage{
counts2fpkm_htseq(filename, mfl_num = c(mfl_number))
}
\arguments{
\item{filename}{The name of the file with htseq counts}

\item{mfl_num}{meanFragmentLength. A numeric vector with mean fragment lengths, which can be calculated using ’CollectInsertSizeMetrics(Picard)’ tool. The length of items should be as the same of columns in read count matrix.}
}
\value{
data frame with fpkm and quartiles
}
\description{
From counts to fpkm and quartiles with htseq files
}
\examples{
\dontrun{
counts2fpkm_htseq (filename, mfl_num = c(mfl_number))
}

}
