% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{day_hour}
\alias{day_hour}
\title{Extract day, hour or day and hour of the moment}
\usage{
day_hour(time1)
}
\arguments{
\item{time1}{day, hour or day_hour}
}
\value{
A character with the information of the moment in different formats to include in names of files
}
\description{
Extract day, hour or day and hour of the moment
}
\examples{
\dontrun {
day_hour ("day")
day_hour ("hour")
day_hour ("day_hour")
}

}
