% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ens2symbol.R
\name{ens2symbol}
\alias{ens2symbol}
\title{Convert ENS to symbols}
\usage{
ens2symbol(
  ens_ids,
  specie = "hsapiens_gene_ensembl",
  attributes_list = c("ensembl_gene_id", "hgnc_symbol", "ensembl_gene_id",
    "chromosome_name", "start_position", "end_position"),
  filter_name = "ensembl_gene_id"
)
}
\arguments{
\item{ens_ids}{a vector with the ensemble ids}

\item{specie}{Specie that we want to get information from}

\item{attributes_list}{attributes that we want to obtain in the last table}

\item{filter_name}{the source information}
}
\value{
a data framewith the symbol, chromosome, start position, end position and length
}
\description{
Search symbols from ENG ids and calculate the length of the gene
}
\examples{
ens2symbol(ens_ids = c("ENSG00000000003"))

}
