% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{filter_format_word}
\alias{filter_format_word}
\title{Extract type of extension}
\usage{
filter_format_word(filename)
}
\arguments{
\item{filename}{Filename that need to be extracted}
}
\value{
last three letters (extension)
}
\description{
Extract type of extension
}
\examples{
\dontrun {
filter_format_word ("testing.doc")
}

}
