% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filtering_mutations.R
\name{filtering_mutations}
\alias{filtering_mutations}
\title{Filtering mutations with certain criteria}
\usage{
filtering_mutations(
  dataframe,
  variant_caller = "Mutect",
  PrimaryFilter = T,
  FuncRefGene = "exonic",
  ExonicFunc = "nonsynonymous_SNV",
  Cob_DNA_tum = 6,
  VAF_tum = 0.05,
  AD_alt_tum = 3,
  AD_alt = 1
)
}
\arguments{
\item{dataframe}{The data frame with all the information}

\item{variant_caller}{Select the variant caller that generated the data frame.}

\item{PrimaryFilter}{If we want to apply PASS filter. Boolean with True by default.}

\item{FuncRefGene}{What function of the Reference Gene we want to apply. Default is exonic.}

\item{ExonicFunc}{What exonic function of the Reference Gene we want to apply. Default value is nonsynonymous_SNV}

\item{Cob_DNA_tum}{Minimum deep of the tumoral DNA (bigger values  will be considered).}

\item{VAF_tum}{Minimum alternative frecuency of the tumor (bigger values  will be considered).}

\item{AD_alt_tum}{Minimum deep of alternative allele of the tumor (bigger values  will be considered).}

\item{AD_alt}{Maximum deep of alternative allele of the healthy tissue (smaller values and selected value will be considered).}
}
\value{
data frame filtered based on the selected criteria
}
\description{
Mutation list will be filtered based on some criteria
}
\examples{
\dontrun{
filtering_mutations (filename)
}

}
