% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_counter2summary.R
\name{four_counter2summary}
\alias{four_counter2summary}
\title{Merge all the counters together}
\usage{
four_counter2summary(
  subread_name,
  cuff_name,
  quant_name,
  htseq_name,
  ngenes = NULL,
  mfl_number = mfl_number,
  export_excel_name = "four_together.xlsx",
  save_intermediate = FALSE
)
}
\arguments{
\item{subread_name}{Data frame with normalized subread cuartiles}

\item{cuff_name}{Data frame with normalized cufflink cuartiles}

\item{quant_name}{Data frame with normalized quant cuartiles}

\item{htseq_name}{Data frame with normalized htseq cuartiles}

\item{ngenes}{Selection of genes that we want to filter}

\item{mfl_number}{Mean length}

\item{export_excel_name}{Name of the final excel name}

\item{save_intermediate}{If we want to save intermediate files}
}
\value{
data frame with all the information merged
}
\description{
Merge all the counters together
}
\examples{
\dontrun{
four_counter2summary (subread_name,cuff_name,quant_name,htseq_name,
                               ngenes = NULL, mfl_number = mfl_number,
                               export_excel_name = "four_together.xlsx",
                               save_intermediate = FALSE)
}

}
