% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm2cuartiles.R
\name{fpkm2cuartiles}
\alias{fpkm2cuartiles}
\title{Fpkm to quartiles}
\usage{
fpkm2cuartiles(fpkm_values)
}
\arguments{
\item{fpkm_values}{}
}
\value{
a vector with quartiles
}
\description{
Fpkm to quartiles
}
\examples{
fpkm2cuartiles_cuff(rnorm(100))

}
