% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fromMultinames2Rows.R
\name{from_multinames_to_rows}
\alias{from_multinames_to_rows}
\title{From Multinames to rows}
\usage{
from_multinames_to_rows(dataset, colu_name)
}
\arguments{
\item{dataset}{The dataset that we want to change}

\item{colu_name}{The name of the column that we want to take into account}
}
\value{
data frame with new rows
}
\description{
Sometimes there are more than one names in the same field. This function helps to separate them in different rows
}
\examples{
\dontrun {
varcall2HowMany (filename)
}

}
