% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_vcf2df.R
\name{mutect38_2df}
\alias{mutect38_2df}
\title{From vcf to excel by mutect38__2df}
\usage{
mutect38_2df(name_vcffile, excel = FALSE, excel_file = "20200306_Mutect.xlsx")
}
\arguments{
\item{name_vcffile}{The name of the vcf file}

\item{excel}{If we want to export to excel (boolean)}

\item{excel_file}{Name of the output excel file.}
}
\value{
data frame with ordered information
}
\description{
From vcf to excel by mutect38__2df
}
\examples{
\dontrun{
mutect38__2df (name_vcffile)
}

}
