% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{select_files_format}
\alias{select_files_format}
\title{Select those elements of a vector with the specified extension}
\usage{
select_files_format(word_vector, formatw = "vcf")
}
\arguments{
\item{word_vector}{Vector with several names}

\item{formatw}{Extension to be filtered}
}
\value{
Select the elements that fill the criteria
}
\description{
Select those elements of a vector with the specified extension
}
\examples{
\dontrun {
select_files_format (c("testing.vcf","bat.vcf","bibi.cod"), formatw = "vcf")
}

}
