% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/from_vcf2df.R
\name{strelka_indels_2_excel}
\alias{strelka_indels_2_excel}
\title{From vcf to excel by Strelka for indels}
\usage{
strelka_indels_2_excel(
  name_vcffile,
  excel = False,
  excel_file = "20200306_strelka_indels.xlsx"
)
}
\arguments{
\item{name_vcffile}{The name of the vcf file}

\item{excel}{If we want to export to excel (boolean)}

\item{excel_file}{Name of the output excel file.}
}
\value{
data frame with ordered information
}
\description{
From vcf to excel by Strelka for indels
}
\examples{
\dontrun{
strelka_2_excel (name_vcffile)
}

}
