% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/varcall2HowMany.R
\name{varcall2HowMany}
\alias{varcall2HowMany}
\title{From excel to How Many varian callers}
\usage{
varcall2HowMany(
  filename,
  chr_pos = 1,
  position = 2,
  gen_name = 3,
  varian_caller = 4,
  VAF = NA,
  others = NULL,
  var_cal_4 = c("mutect38", "somaticsniper", "strelka", "varscan")
)
}
\arguments{
\item{filename}{The name of the excel file with the mutations with choromosome, position, gen name and variant caller}

\item{chr_pos}{Position of the chromosome name in the data frame}

\item{position}{Position of the mega position in the data frame}

\item{gen_name}{Position of the Gene name in the data frame}

\item{VAF}{Position of the VAF name in the data frame (if is not present, NA will be presented)}

\item{others}{Position of the other variables in the data frame that will be merged}

\item{var_cal_4}{Names of the four variant caller to calculate the How many field}
}
\value{
data frame with How many variant callers per mutation
}
\description{
From an excel with mutation detected in each varian caller, the function convert the data in a table with information about in how many variant callers appeared the mutation.
}
\examples{
\dontrun{
varcall2HowMany (filename)
}

}
