% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_counter2summary.R
\name{four_counter2summary}
\alias{four_counter2summary}
\title{Merge all the counters together}
\usage{
four_counter2summary(
  subread_name,
  cuff_name,
  quant_name,
  htseq_name,
  ngenes,
  mfl_number = mfl_number,
  export_excel_name = "four_together.xlsx",
  save_intermediate = FALSE
)
}
\arguments{
\item{subread_raw_file}{Data frame with normalized subread cuartiles}

\item{cufflink_raw_file}{Data frame with normalized cufflink cuartiles}

\item{htseq_raw_file}{Data frame with normalized htseq cuartiles}

\item{quant_raw_file}{Data frame with normalized quant cuartiles}
}
\value{
data frame with all the information merged
}
\description{
Merge all the counters together
}
\examples{
\dontrun {
four_counter2summary (subread_name,cuff_name,quant_name,htseq_name,
                               ngenes, mfl_number = mfl_number, export_excel_name = "four_together.xlsx", save_intermediate = FALSE)
}

}
